Move ScrtionList from WhitePaper.
